Don't expect reflection information about incomplete entities.
When we have a forward declaration this means that the entity is opaque to the
compiler. Asking further details generally should work, however, there is a lot
of gray zones. For example, if we ask if a forward declared entity is abstract,
there is no yes/no answer.

The C++ type system has the notion of complete/incomplete types where we can
do our best with the little information we have. This patch introduces this
into the way we generate proxy calls for the C++ enities. This needs quite a bit
more refactoring.